Number of doctorates versus arcade revenue in the United States

A dataset containing the yearly number of doctorates awarded in computer science and the total revenue generated by arcades in the United States for the period 2000-2009.

arcade

Format

A data frame with 10 rows and 3 variables:

year

Year of the observation.

doctorates

Number of doctorates awarded in computer science.

revenue

Total revenue generated by arcades (in billions of dollars).

Source

Spurious Correlations (Vigen 2015)

Examples

arcade
#> # A tibble: 10 × 3
#>     year doctorates revenue
#>    <int>      <dbl>   <dbl>
#>  1  2000        861    1.20
#>  2  2001        830    1.18
#>  3  2002        809    1.27
#>  4  2003        867    1.24
#>  5  2004        948    1.31
#>  6  2005       1129    1.44
#>  7  2006       1453    1.60
#>  8  2007       1656    1.65
#>  9  2008       1787    1.80
#> 10  2009       1611    1.73